fix: refactor RoutedToPerPeriod to use Insights Routing Service
