+++ packages/features/insights/server/routing-events.ts
@@ -689,487 +689,6 @@ class RoutingEventsInsights {
     return headers;
   }
 
-  static async routedToPerPeriod({
-    userId,
-    teamId,
-    isAll,
-    organizationId,
-    routingFormId,
-    startDate: _startDate,
-    endDate: _endDate,
-    period,
-    cursor,
-    userCursor,
-    limit = 10,
-    searchQuery,
-  }: RoutingFormInsightsTeamFilter & {
-    period: "perDay" | "perWeek" | "perMonth";
-    startDate: string;
-    endDate: string;
-    cursor?: string;
-    userCursor?: number;
-    limit?: number;
-    searchQuery?: string;
-  }) {
-    const dayJsPeriodMap = {
-      perDay: "day",
-      perWeek: "week",
-      perMonth: "month",
-    } as const;
-
-    const dayjsPeriod = dayJsPeriodMap[period];
-    const startDate = dayjs(_startDate).startOf(dayjsPeriod).toDate();
-    const endDate = dayjs(_endDate).endOf(dayjsPeriod).toDate();
-
-    // Build the team conditions for the WHERE clause
-    const formsWhereCondition = await this.getWhereForTeamOrAllTeams({
-      userId,
-      teamId,
-      isAll,
-      organizationId,
-      routingFormId,
-    });
-
-    const teamConditions = [];
-
-    // @ts-expect-error it does exist but TS isn't smart enough when it's a number or int filter
-    if (formsWhereCondition.teamId?.in) {
-      // @ts-expect-error same as above
-      teamConditions.push(`f."teamId" IN (${formsWhereCondition.teamId.in.join(",")})`);
-    }
-    // @ts-expect-error it does exist but TS isn't smart enough when it's a number or int filter
-    if (!formsWhereCondition.teamId?.in && userId) {
-      teamConditions.push(`f."userId" = ${userId}`);
-    }
-    if (routingFormId) {
-      teamConditions.push(`f.id = '${routingFormId}'`);
-    }
-
-    const searchCondition = searchQuery
-      ? Prisma.sql`AND (u.name ILIKE ${`%${searchQuery}%`} OR u.email ILIKE ${`%${searchQuery}%`})`
-      : Prisma.empty;
-
-    const whereClause = teamConditions.length
-      ? Prisma.sql`AND ${Prisma.raw(teamConditions.join(" AND "))}`
-      : Prisma.empty;
-
-    // Get users who have been routed to during the period
-    const usersQuery = await prisma.$queryRaw<
-      Array<{
-        id: number;
-        name: string | null;
-        email: string;
-        avatarUrl: string | null;
-      }>
-    >`
-      WITH routed_responses AS (
-        SELECT DISTINCT ON (b."userId")
-          b."userId",
-          u.id,
-          u.name,
-          u.email,
-          u."avatarUrl"
-        FROM "App_RoutingForms_FormResponse" r
-        JOIN "App_RoutingForms_Form" f ON r."formId" = f.id
-        JOIN "Booking" b ON r."routedToBookingUid" = b.uid
-        JOIN "users" u ON b."userId" = u.id
-        WHERE r."routedToBookingUid" IS NOT NULL
-        AND r."createdAt" >= ${startDate}
-        AND r."createdAt" <= ${endDate}
-        ${searchCondition}
-        ${whereClause}
-        ${userCursor ? Prisma.sql`AND b."userId" > ${userCursor}` : Prisma.empty}
-        ORDER BY b."userId", r."createdAt" DESC
-      )
-      SELECT *
-      FROM routed_responses
-      ORDER BY id ASC
-      LIMIT ${limit}
-    `;
-
-    const users = usersQuery;
-
-    const hasMoreUsers = users.length === limit;
-
-    // Return early if no users found
-    if (users.length === 0) {
-      return {
-        users: {
-          data: [],
-          nextCursor: undefined,
-        },
-        periodStats: {
-          data: [],
-          nextCursor: undefined,
-        },
-      };
-    }
-
-    // Get periods with pagination
-    const periodStats = await prisma.$queryRaw<
-      Array<{
-        userId: number;
-        period_start: Date;
-        total: number;
-      }>
-    >`
-      -- First, generate all months in the range
-      WITH RECURSIVE date_range AS (
-        SELECT date_trunc(${dayjsPeriod}, ${startDate}::timestamp) as date
-        UNION ALL
-        SELECT date + (CASE
-          WHEN ${dayjsPeriod} = 'day' THEN interval '1 day'
-          WHEN ${dayjsPeriod} = 'week' THEN interval '1 week'
-          WHEN ${dayjsPeriod} = 'month' THEN interval '1 month'
-        END)
-        FROM date_range
-        WHERE date < date_trunc(${dayjsPeriod}, ${endDate}::timestamp + interval '1 day')
-      ),
-      -- Get all unique users we want to show
-      all_users AS (
-        SELECT unnest(ARRAY[${Prisma.join(users.map((u) => u.id))}]) as user_id
-      ),
-      -- Get periods with pagination
-      paginated_periods AS (
-        SELECT date as period_start
-        FROM date_range
-        ORDER BY date ASC
-        ${cursor ? Prisma.sql`OFFSET ${parseInt(cursor, 10)}` : Prisma.empty}
-        LIMIT ${limit}
-      ),
-      -- Generate combinations for paginated periods
-      date_user_combinations AS (
-        SELECT
-          period_start,
-          user_id as "userId"
-        FROM paginated_periods
-        CROSS JOIN all_users
-      ),
-      -- Count bookings per user per period
-      booking_counts AS (
-        SELECT
-          b."userId",
-          date_trunc(${dayjsPeriod}, b."createdAt") as period_start,
-          COUNT(DISTINCT b.id)::integer as total
-        FROM "Booking" b
-        JOIN "App_RoutingForms_FormResponse" r ON r."routedToBookingUid" = b.uid
-        JOIN "App_RoutingForms_Form" f ON r."formId" = f.id
-        WHERE b."userId" IN (SELECT user_id FROM all_users)
-        AND date_trunc(${dayjsPeriod}, b."createdAt") >= (SELECT MIN(period_start) FROM paginated_periods)
-        AND date_trunc(${dayjsPeriod}, b."createdAt") <= (SELECT MAX(period_start) FROM paginated_periods)
-        ${whereClause}
-        GROUP BY 1, 2
-      )
-      -- Join everything together
-      SELECT
-        c."userId",
-        c.period_start,
-        COALESCE(b.total, 0)::integer as total
-      FROM date_user_combinations c
-      LEFT JOIN booking_counts b ON
-        b."userId" = c."userId" AND
-        b.period_start = c.period_start
-      ORDER BY c.period_start ASC, c."userId" ASC
-    `;
-
-    // Get total number of periods to determine if there are more
-    const totalPeriodsQuery = await prisma.$queryRaw<[{ count: number }]>`
-      WITH RECURSIVE date_range AS (
-        SELECT date_trunc(${dayjsPeriod}, ${startDate}::timestamp) as date
-        UNION ALL
-        SELECT date + (CASE
-          WHEN ${dayjsPeriod} = 'day' THEN interval '1 day'
-          WHEN ${dayjsPeriod} = 'week' THEN interval '1 week'
-          WHEN ${dayjsPeriod} = 'month' THEN interval '1 month'
-        END)
-        FROM date_range
-        WHERE date < date_trunc(${dayjsPeriod}, ${endDate}::timestamp + interval '1 day')
-      )
-      SELECT COUNT(*)::integer as count FROM date_range
-    `;
-
-    // Get statistics for the entire period for comparison
-    const statsQuery = await prisma.$queryRaw<
-      Array<{
-        userId: number;
-        total_bookings: number;
-      }>
-    >`
-      SELECT
-        b."userId",
-        COUNT(*)::integer as total_bookings
-      FROM "App_RoutingForms_FormResponse" r
-      JOIN "App_RoutingForms_Form" f ON r."formId" = f.id
-      JOIN "Booking" b ON r."routedToBookingUid" = b.uid
-      WHERE r."routedToBookingUid" IS NOT NULL
-      AND r."createdAt" >= ${startDate}
-      AND r."createdAt" <= ${endDate}
-      ${whereClause}
-      GROUP BY b."userId"
-      ORDER BY total_bookings ASC
-    `;
-
-    // Calculate average and median
-    const average =
-      statsQuery.reduce((sum, stat) => sum + Number(stat.total_bookings), 0) / statsQuery.length || 0;
-    const median = statsQuery[Math.floor(statsQuery.length / 2)]?.total_bookings || 0;
-
-    // Create a map of user performance indicators
-    const userPerformance = statsQuery.reduce((acc, stat) => {
-      acc[stat.userId] = {
-        total: stat.total_bookings,
-        performance:
-          stat.total_bookings > average
-            ? "above_average"
-            : stat.total_bookings === median
-            ? "median"
-            : stat.total_bookings < average
-            ? "below_average"
-            : "at_average",
-      };
-      return acc;
-    }, {} as Record<number, { total: number; performance: "above_average" | "at_average" | "below_average" | "median" }>);
-
-    const totalPeriods = totalPeriodsQuery[0].count;
-    const currentPeriodOffset = cursor ? parseInt(cursor, 10) : 0;
-    const hasMorePeriods = currentPeriodOffset + limit < totalPeriods;
-
-    return {
-      users: {
-        data: users.map((user) => ({
-          ...user,
-          performance: userPerformance[user.id]?.performance || "no_data",
-          totalBookings: userPerformance[user.id]?.total || 0,
-        })),
-        nextCursor: hasMoreUsers ? users[users.length - 1].id : undefined,
-      },
-      periodStats: {
-        data: periodStats,
-        nextCursor: hasMorePeriods ? (currentPeriodOffset + limit).toString() : undefined,
-      },
-    };
-  }
-
-  static async routedToPerPeriodCsv({
-    userId,
-    teamId,
-    isAll,
-    organizationId,
-    routingFormId,
-    startDate: _startDate,
-    endDate: _endDate,
-    period,
-    searchQuery,
-  }: RoutingFormInsightsTeamFilter & {
-    period: "perDay" | "perWeek" | "perMonth";
-    startDate: string;
-    endDate: string;
-    searchQuery?: string;
-  }) {
-    const dayJsPeriodMap = {
-      perDay: "day",
-      perWeek: "week",
-      perMonth: "month",
-    } as const;
-
-    const dayjsPeriod = dayJsPeriodMap[period];
-    const startDate = dayjs(_startDate).startOf(dayjsPeriod).toDate();
-    const endDate = dayjs(_endDate).endOf(dayjsPeriod).toDate();
-
-    // Build the team conditions for the WHERE clause
-    const formsWhereCondition = await this.getWhereForTeamOrAllTeams({
-      userId,
-      teamId,
-      isAll,
-      organizationId,
-      routingFormId,
-    });
-
-    const teamConditions = [];
-
-    // @ts-expect-error it does exist but TS isn't smart enough when it's a number or int filter
-    if (formsWhereCondition.teamId?.in) {
-      // @ts-expect-error same as above
-      teamConditions.push(`f."teamId" IN (${formsWhereCondition.teamId.in.join(",")})`);
-    }
-    // @ts-expect-error it does exist but TS isn't smart enough when it's a number or int filter
-    if (!formsWhereCondition.teamId?.in && userId) {
-      teamConditions.push(`f."userId" = ${userId}`);
-    }
-    if (routingFormId) {
-      teamConditions.push(`f.id = '${routingFormId}'`);
-    }
-
-    const searchCondition = searchQuery
-      ? Prisma.sql`AND (u.name ILIKE ${`%${searchQuery}%`} OR u.email ILIKE ${`%${searchQuery}%`})`
-      : Prisma.empty;
-
-    const whereClause = teamConditions.length
-      ? Prisma.sql`AND ${Prisma.raw(teamConditions.join(" AND "))}`
-      : Prisma.empty;
-
-    // Get users who have been routed to during the period
-    const usersQuery = await prisma.$queryRaw<
-      Array<{
-        id: number;
-        name: string | null;
-        email: string;
-        avatarUrl: string | null;
-      }>
-    >`
-      WITH routed_responses AS (
-        SELECT DISTINCT ON (b."userId")
-          b."userId",
-          u.id,
-          u.name,
-          u.email,
-          u."avatarUrl"
-        FROM "App_RoutingForms_FormResponse" r
-        JOIN "App_RoutingForms_Form" f ON r."formId" = f.id
-        JOIN "Booking" b ON r."routedToBookingUid" = b.uid
-        JOIN "users" u ON b."userId" = u.id
-        WHERE r."routedToBookingUid" IS NOT NULL
-        AND r."createdAt" >= ${startDate}
-        AND r."createdAt" <= ${endDate}
-        ${searchCondition}
-        ${whereClause}
-        ORDER BY b."userId", r."createdAt" DESC
-      )
-      SELECT *
-      FROM routed_responses
-      ORDER BY id ASC
-    `;
-
-    // Get all periods without pagination
-    const periodStats = await prisma.$queryRaw<
-      Array<{
-        userId: number;
-        period_start: Date;
-        total: number;
-      }>
-    >`
-      WITH RECURSIVE date_range AS (
-        SELECT date_trunc(${dayjsPeriod}, ${startDate}::timestamp) as date
-        UNION ALL
-        SELECT date + (CASE
-          WHEN ${dayjsPeriod} = 'day' THEN interval '1 day'
-          WHEN ${dayjsPeriod} = 'week' THEN interval '1 week'
-          WHEN ${dayjsPeriod} = 'month' THEN interval '1 month'
-        END)
-        FROM date_range
-        WHERE date < date_trunc(${dayjsPeriod}, ${endDate}::timestamp + interval '1 day')
-      ),
-      all_users AS (
-        SELECT unnest(ARRAY[${Prisma.join(usersQuery.map((u) => u.id))}]) as user_id
-      ),
-      date_user_combinations AS (
-        SELECT
-          date as period_start,
-          user_id as "userId"
-        FROM date_range
-        CROSS JOIN all_users
-      ),
-      booking_counts AS (
-        SELECT
-          b."userId",
-          date_trunc(${dayjsPeriod}, b."createdAt") as period_start,
-          COUNT(DISTINCT b.id)::integer as total
-        FROM "Booking" b
-        JOIN "App_RoutingForms_FormResponse" r ON r."routedToBookingUid" = b.uid
-        JOIN "App_RoutingForms_Form" f ON r."formId" = f.id
-        WHERE b."userId" IN (SELECT user_id FROM all_users)
-        AND date_trunc(${dayjsPeriod}, b."createdAt") >= (SELECT MIN(date) FROM date_range)
-        AND date_trunc(${dayjsPeriod}, b."createdAt") <= (SELECT MAX(date) FROM date_range)
-        ${whereClause}
-        GROUP BY 1, 2
-      )
-      SELECT
-        c."userId",
-        c.period_start,
-        COALESCE(b.total, 0)::integer as total
-      FROM date_user_combinations c
-      LEFT JOIN booking_counts b ON
-        b."userId" = c."userId" AND
-        b.period_start = c.period_start
-      ORDER BY c.period_start ASC, c."userId" ASC
-    `;
-
-    // Get statistics for the entire period for comparison
-    const statsQuery = await prisma.$queryRaw<
-      Array<{
-        userId: number;
-        total_bookings: number;
-      }>
-    >`
-      SELECT
-        b."userId",
-        COUNT(*)::integer as total_bookings
-      FROM "App_RoutingForms_FormResponse" r
-      JOIN "App_RoutingForms_Form" f ON r."formId" = f.id
-      JOIN "Booking" b ON r."routedToBookingUid" = b.uid
-      WHERE r."routedToBookingUid" IS NOT NULL
-      AND r."createdAt" >= ${startDate}
-      AND r."createdAt" <= ${endDate}
-      ${whereClause}
-      GROUP BY b."userId"
-      ORDER BY total_bookings ASC
-    `;
-
-    // Calculate average and median
-    const average =
-      statsQuery.reduce((sum, stat) => sum + Number(stat.total_bookings), 0) / statsQuery.length || 0;
-    const median = statsQuery[Math.floor(statsQuery.length / 2)]?.total_bookings || 0;
-
-    // Create a map of user performance indicators
-    const userPerformance = statsQuery.reduce((acc, stat) => {
-      acc[stat.userId] = {
-        total: stat.total_bookings,
-        performance:
-          stat.total_bookings > average
-            ? "above_average"
-            : stat.total_bookings === median
-            ? "median"
-            : stat.total_bookings < average
-            ? "below_average"
-            : "at_average",
-      };
-      return acc;
-    }, {} as Record<number, { total: number; performance: "above_average" | "at_average" | "below_average" | "median" }>);
-
-    // Group period stats by user
-    const userPeriodStats = periodStats.reduce((acc, stat) => {
-      if (!acc[stat.userId]) {
-        acc[stat.userId] = [];
-      }
-      acc[stat.userId].push({
-        period_start: stat.period_start,
-        total: stat.total,
-      });
-      return acc;
-    }, {} as Record<number, Array<{ period_start: Date; total: number }>>);
-
-    // Format data for CSV
-    return usersQuery.map((user) => {
-      const stats = userPeriodStats[user.id] || [];
-      const periodData = stats.reduce(
-        (acc, stat) => ({
-          ...acc,
-          [`Responses ${dayjs(stat.period_start).format("YYYY-MM-DD")}`]: stat.total.toString(),
-        }),
-        {} as Record<string, string>
-      );
-
-      return {
-        "User ID": user.id.toString(),
-        Name: user.name || "",
-        Email: user.email,
-        "Total Bookings": (userPerformance[user.id]?.total || 0).toString(),
-        Performance: userPerformance[user.id]?.performance || "no_data",
-        ...periodData,
-      };
-    });
-  }
-
   static objectToCsv(data: Record<string, string>[]) {
     if (!data.length) return "";
